Se incorpora dialogo de nueva tarea y se elimina editText en pntalla principal

+++ app/src/main/java/com/example/listadetareas/ui/MainActivity.kt
@@ -1,21 +1,26 @@
 package com.example.listadetareas.ui
 
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
 import android.widget.Toast
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView.*
+import com.example.listadetareas.R
 import com.example.listadetareas.databinding.ActivityMainBinding
+import com.example.listadetareas.databinding.NewTaskBinding
 import com.example.listadetareas.ui.adapter.OnClickListener
 import com.example.listadetareas.ui.adapter.TasksAdapter
 import com.example.listadetareas.ui.model.TaskModel
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+
 
 class MainActivity : AppCompatActivity(), OnClickListener {
 
     private lateinit var mBinding: ActivityMainBinding
-    private var tasks: MutableList<TaskModel> = mutableListOf()
-
     private lateinit var mAdapter: TasksAdapter
 
     //LinearLayout
@@ -30,36 +35,20 @@ class MainActivity : AppCompatActivity(), OnClickListener {
         mBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(mBinding.root)
 
-        //Establezco la configuración de observadores del ViewModel
-        initViewModel()
         // Inicializar el recyclerView
         initRecyclerView()
+        //Establezco la configuración de observadores del ViewModel
+        initViewModel()
 
         // Manejar el clic en el botón de agregar tarea
         mBinding.btnAddTask.setOnClickListener {
-            // Obtener el texto ingresado en el campo de texto
-            val taskName = mBinding.etNameTask.text.toString().trim()
-
-            // Verificar si el nombre no está vacío
-            if (taskName.isNotEmpty()) {
-                // Agregar la tarea a la lista
-                mainViewModel.addTask(
-                    TaskModel(
-                        id = System.currentTimeMillis(),
-                        name = mBinding.etNameTask.text.toString().trim(),
-                        description = "",
-                        finish = false
-                    )
-                )
-                // Limpiar el campo de texto
-                mBinding.etNameTask.text.clear()
-            }
+            showDialog()
         }
     }
 
     private fun initRecyclerView() {
         //una unica columna
-        mAdapter = TasksAdapter(tasks, this)
+        mAdapter = TasksAdapter(mutableListOf(), this)
         mLinearLayout = LinearLayoutManager(this)
         //Grid
         // mGridLayout = GridLayoutManager(this, spanCount)
@@ -72,33 +61,74 @@ class MainActivity : AppCompatActivity(), OnClickListener {
 
     private fun initViewModel() {
 
+        mainViewModel.getAllTasks()
+
         mainViewModel.tasks.observe(this) { tasksList ->
             //Si el listado de tareas cambia, se actualiza el RV
-
-            if (tasksList.isNotEmpty())
-                mAdapter = TasksAdapter(tasksList, this)
-                mBinding.recyclerView.adapter = mAdapter
-                mBinding.recyclerView.adapter?.notifyItemChanged(tasksList.size - 1)
+            mAdapter = TasksAdapter(tasksList, this)
+            mBinding.recyclerView.adapter = mAdapter
+            mBinding.recyclerView.adapter?.notifyItemChanged(tasksList.size - 1)
         }
-
     }
-
-
-    //Esta funcion se llama desde el OnCreate
-
 
     override fun onDeleteTask(taskModel: TaskModel) {
         // Eliminar la tarea de la lista
-        mainViewModel.onDeleteTask(taskModel)
-        Toast.makeText(this, "Item eliminado", Toast.LENGTH_SHORT).show()
+        MaterialAlertDialogBuilder(this)
+            .setTitle(getString(R.string.text_title_alert_delete))
+            .setMessage(getString(R.string.text_explain_delete).plus("${taskModel.name}?"))
+            .setNegativeButton("Descartar") { _, _ ->
+                // Respond to neutral button press
+            }
+            .setPositiveButton("Eliminar") { _, _ ->
+                mainViewModel.onDeleteTask(taskModel)
+                Toast.makeText(this, "Item eliminado", Toast.LENGTH_SHORT).show()
+            }
+            .show()
     }
 
     override fun onUpdateTask(taskModel: TaskModel) {
+        //Actualizar la tarea seleccionada
         mainViewModel.onUpdateTask(taskModel)
         Toast.makeText(this, "Item actualizado", Toast.LENGTH_SHORT).show()
     }
 
     override fun onClickItem(taskModel: TaskModel) {
         Toast.makeText(this, "Has pulsado la tarea: ${taskModel.name}", Toast.LENGTH_SHORT).show()
+    }
+
+    private fun showDialog() {
+
+        val dialogBinding = NewTaskBinding.inflate(LayoutInflater.from(this))
+        val dialogView: View = dialogBinding.root
+
+        val dialogBuilder: AlertDialog.Builder = AlertDialog.Builder(this)
+            .setView(dialogView)
+            .setTitle("Añadir Tarea")
+            .setCancelable(false)
+
+        val dialog: AlertDialog = dialogBuilder.create()
+        dialog.show()
+
+        dialogBinding.btnCancelTask.setOnClickListener {
+            dialog.dismiss()
+        }
+
+        dialogBinding.btnAddTaskDialog.setOnClickListener {
+
+            val name = dialogBinding.etNameTaskDialog.text.toString()
+            val description = dialogBinding.etDescriptionDialog.text.toString()
+
+            if (name.isNotEmpty()) {
+                // Agregar la tarea a la lista
+                mainViewModel.addTask(
+                    TaskModel(
+                        name = name.trim(),
+                        description = description.trim(),
+                        finish = false
+                    )
+                )
+                dialog.dismiss()
+            }
+        }
     }
 }
